Android: ensure the preferred ABI is parsed early on under QtLoader ctor
Ensure the preferred ABI is obtained separately early on before doing
anything else. At the moment it is assigned a value when it's first
accessed by some call that needs it and some calls might not work if
it's null, so that way it can easily cause unexpected behavior.

Now, that we should be guaranteed the perferred ABI to be resolved
first, and the libs list is not going to change later during the
lifetime of the app, we can populate only one ArrayList of libs
instead of a Map of libs for each ABI.

Task-number: QTBUG-132211
Task-number: QTBUG-120467
Task-number: QTBUG-70114
Task-number: QTBUG-86314
Pick-to: 6.8
Change-Id: I89a6d3aef5c2df0134579e3d12c04e4c5d7e021a
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>
(cherry picked from commit 952a785003c9955c3536cf9cacfd502efc84320f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 1c521c65cc4790ee9992494d345700136f6ad99c)