QDirListing: check for '.' and '..' after the checking the name filters
If we check for '.' and '..' before checking the name filters one or
both of them may match if IncludeDotAndDotDot is set, but should be
rejected based on the name filters.

This matches the existing behavior of QDir/QDirIterator:
- With QDir::Filters, '.' and '..' are listed by default unless
  QDir::NoDot and QDir::NoDotDot are set; so checking NoDot and NoDotDot
  before chekcing the name filters, still means the name filters are
  applied if '.' and '..' are listed. See matchesLegacyFilters().
- With QDirListing '.' and '..' aren't listed unless InlcudeDotAndDotDot
  is set. So the name filters have to be checked before checking for the
  special entrie '.'  and '..'.

This should ease the porting to QDirListing.

This is slightly less efficient when using QDirListing directly (not via
QDir::entryList()); but if you want efficiency, you shouldn't be using
name filters which use regex matching, to begin with.

[ChangeLog][QtCore][QDirListing] Now listing '.' and '..' special
entries respects any set name filters. Previously if IncludeDotAndDotDot
was set, '.' and '..' would always be listed.

Pick-to: 6.9 6.8
Change-Id: I026e1e44431515ab2cc3545f993580dbf010ac9b
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 5f012d8b256475fda2f4182245336390387972fc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>