QLocale/ICU: don't take addresses of ICU functions
Taking addresses of library functions is usually not a good idea. The
C++ standard makes it UB for its own symbols, and while I didn't check
whether ICU explicitly allows it, the answer would probably be that
no-one ever thought about this...

Port from passing pointers to library functions to passing an enum
{Upper, Lower}.

As a drive-by, wrap most of the functions in a namespace
QtIcuPrivate. I did this initially to keep the enum name manageable,
but it looks kinda stupid to have just the enum in the namespace and
not the function taking it as an argument now.

Task-number: QTBUG-138583
Pick-to: 6.9 6.8 6.5
Change-Id: I31b21737468bf05f52f695279dd071117420736e
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 6f4d64b21ec90a792c7c8308a6615ca87b9ea6eb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>