windeployqt: Fix handling of special deployment of ICU libraries
ICU libraries are not Qt modules and thus should not be added to
dependentQtLibs. Adding it to dependentQtLibs will cause the "module
magic" which will result in an error because it cannot be found in our
module map.

To avoid this magic (and the error message), the ICU libraries should be
added to deployedQtLibraries directly.

Fixes: QTBUG-138474
Change-Id: Ifef90066a74ae277ca4ef1775ce1b720ecbb6e40
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit c93cbe39b408c1e8b0c0e7d8152f01d06b07fd2a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>