QMacStyle: tweak pushbutton's bezel geometry
In macOS Tahoe the bezel is clipped on left/right sides and this makes push button look square.
In previous versions of the macOS button bezel was rendered by the AppKit smaller than
the rectangle we want to draw it with, and we worked this problem around by adjusting
the rectangle, shifting it to the left and adding to its width
(and the resulting bezel thus was fitting within the clip rectangle).
In macOS Tahoe these adjustments not needed anymore - the button's bezel
is rendered exactly within the rectangle we ask to draw with (and this also means
within the clipping rectangle).

Pick-to: 6.8 6.5
Task-number: QTBUG-138699
Change-Id: I7f203808c24276c5ee982b7e412108681bc57733
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 75cd031c208b257fc42e58069ae23470f11f132d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 335b30b891956fa7ee13c3c55edb7b4843c6b756)