Http: Remove incomplete cache-objects on destruction
Most commonly this would be a file, by way of the network disk cache.

The switch to use QSaveFile in the network disk cache in
e5f295c8a458dcd336f7cf3768ca62aded69e659 exposed the issue because
the leftover file objects were no longer deleted during
the cache trimming, so the handles were left open even if we
deleted the file.

The real issue, however, is that we did not notify the cache
that we were not going to complete the cache object, thus closing
the file object and deleting the uncommitted file.

That issue was probably introduced as part of
6f25051536c1636688a0a0939196007aac34676d.

Pick-to: 6.9 6.8 6.5
Fixes: QTBUG-135641
Change-Id: I4c185e4d6e44029e221e69e2ef7135b3710f1069
Reviewed-by: Jonas Kvinge <jonas@jkvinge.net>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit 27f939da8fa08c75376fd0b009d83aafc8e04bc9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>