wayland: Fix cursor state management
setMouseCursor was used exclusively for CSDs, we shouldn't be storing it
in the cursor management. Otherwise when the cursor is restored we
restored when we move out of decoration we restore the application set
cursor. We should just apply the cursor directly as it takes precedent.

In addition the handling of unsetting a cursor back to default was not
implemented. We need to unset the application stored cache and always
update the server state.

Updates 555bda173621e18ae15117677ed3f314cee4a6bb

Fixes: QTBUG-138542
Change-Id: I6e701471f6b0500e6f301a0c8dcb917b4f4439b1
Reviewed-by: Błażej Szczygieł <mumei6102@gmail.com>
Reviewed-by: David Redondo <qt@david-redondo.de>
(cherry picked from commit f06fd493f5fa94758161f5d3fbcf6f49dffc7c0b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>