qvsnprintf: fail if the result size doesn't fit into an int
The old code silently truncated the return value of the fallback
implementation. This is dangerous, as the caller might rely on the
truncated value to confirm that everything was written when it
actually wasn't, incl. trusting the return value to allocate a buffer
of a size dependent on the result and then strcpy()ing the data over
(which would overflow the target buffer, since the NUL byte is
actually at `n` and not at `result`, as the function post-condition
would require for a result < n.

POSIX specifies¹ that errno be set to EOVERFLOW and -1 be returned, so
do that.

As a drive-by, replace use of qMin() with std::min() (which statically
asserts that both arguments are of the same type, unlike qMin(), which
performs integral promotion on the arguments).

Use qt_saturate() for now, because 6.5 doesn't have
q26::saturate_cast(). A follow-up patch will clean this up for
6.8...dev.

[ChangeLog][QtCore][q(v)snprintf] Fixed a bug that would cause
q(v)snprintf() to report success on some platforms, even though the
result was truncated.

Amends the port of Qt containers from int to qsizetype. Didn't check
for a particular commit, because a) it's probably several that apply
and b) they are so much older than the currently active branches that
I didn't feel I need to be more specific than, essentially, "Qt 6.0".

¹ https://pubs.opengroup.org/onlinepubs/9699919799/functions/fprintf.html

Fixes: QTBUG-138860
Pick-to: 6.9 6.8 6.5
Change-Id: I6de8a99c238539a535462e964a0645bd0c704fff
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit ed3589374523ea5cef3ff572eab94a137629025c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>