QRM: reuse read/writeRole helpers in itemData/setItemData
There's no reason to duplicate the logic and specialization for QObject,
we do all of that in the helpers already. The helpers also fall-back to
Qt::DisplayRole when a property-lookup for Qt::EditRole fails, which we
also want for itemRole - if roleNames() includes that role, then it
should be part of the map of values returned.

This requires a small adjustment to the test, which already takes into
account that Qt::EditRole is faked in our implementation anyway.

Change-Id: Id3d77cf7e83b4220ab2ac8aae7bba2a9de7d16d8
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 6b92bc168717cd28ad2f0a1c62d4d4d099b8a35f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>