HTTP: don't send content-length: 0
If a request is made with a valid QIODevice that happens to be empty
then we end up setting the Content-Length header to 0.

This is discouraged according to the RFC9110, specifically for methods
that do not usually expect a body.

Pick-to: 6.9 6.8
Fixes: QTBUG-138848
Change-Id: I618220008ed9f739fed96e5fd7630a0afd5cdaaf
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit bb40f641f2f80ff5a01bf49d50470d6868cbf701)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>