tst_qnetworkreply_local: shorten the local socket names
It seems to be failing with some test cases on some Linux systems.
It doesn't actually fail to listen on the server-side, but it seems the
client-side socket instead fails to connect.

Instead of using the full test case names and process id, hash it
together and use the HEX of that instead for the local socket
identifier.

Change in 6.10 cherry-pick:
 - Resolved conflict due to 0b3d7b9e60480f0721019a0fe4b35ecd8f52d9a8
   not being cherry-picked.

Pick-to: 6.9 6.8
Change-Id: Ibcd0f6af29652d68470596efe84fba954cda57ce
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit ea5471e3b84a1ccaa5b02618d91bec6130efc3b1)