QCalendarLocale: move switch over QLocale::FormatType into accessors
Before, each user of QCalendarLocale was performing the switch by
themselves. Even though there was only one user each of month and
standalone month, their switch statements were buried in larger
functions and contained a default case label with the return from the
surrounding function. The default was necessary, because running off
the end of the switch was normal operation, but it destroys -Wswitch.

Instead of wrapping the switch in a lambda, realize that these
switches were also the only users of the fine-grained accessors of
QCalendarLocale and move the would-be lambda there as a member
function. This means we now have functions containing only one switch
each, so we can return from each case label and falling off the end of
the switch statement is unreachable. This removes the need for a
default case label, re-enabling -Wswitch. It also allows to mark the
DataRange objects const in the callers.

Found as part of fixing Coverity issue 425487.

Amends aa8393c94fea01a4806b204fd3aa343a4e90666b.

Coverity-Id: 425487
Pick-to: 6.9 6.8 6.5
Change-Id: If1dd16f11b96372f00dc4469fd6b12f835bc7877
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit 4bef66d2dcebda406d2412d168d5b9f07c2b3c37)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>