QLocale: avoid QList self-assignment
QStringList::replaceInStrings() returns a reference to *this, so the
re-assignment to *this is unnecessary.

If one nevertheless does it, the RHS binds to the copy assignment
operator, which, for QList, is copy-and-swap, so it's
self-assignment-safe, but ups and downs the ref-count needlessly.

Remove the assignment.

Amends 4431ffa8b90f4d089f98544e8c2eeac6cfc7abb2.

Pick-to: 6.9 6.8
Change-Id: I2e44b4ac302b1d9193d29e41bdbfc775d4682457
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit e1ecc3ae11e01f6a6601b921efb588e13b8baf3b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>