QLocale: fix another Coverity complaint about missing moves
Here, Coverity is complaining that 'prior' is copied into the lambda
when it should be moved.

But 'prior' is marked as const, so to move it, we'd first need to make
it non-const, turning a constant into yet another variable in this
already huge function.

Instead, capture 'prior' by reference. Since 'prior' is const, it's
essentially the same as by-value, sans the copy that Coverity was
complaining about.

Amends b6122866962880251fa8a82856d1fb9db43900d6.

According to my reading of the Coverity database, this was the last
issue in files matching */qlocale*, so, provided the previous patches
worked as expected, QLocale is now Coverity-clean.

Coverity-Id: 474124
Pick-to: 6.9
Change-Id: Ide4dd9605c54e3aec7430146a909853c5395238f
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit fc3f4b1b0ec1180b3a02f50078f9af396697b7a0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>