CMake: fix the reduce_relocations test
This test was apparently never properly ported from qmake, because it
forgot to check that the linker supports the flag that it meant to use.
Fix it by checking for both flags that we need to support the feature,
-Bsymbolic-functions and --dynamic-list.

This also fixes the incorrect GCC checks, since this is not a compiler
feature, but a linker one.

Change-Id: I92f6bb1e213e335b5c94fffd3ea509c53b847ee6
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit 1f49b5076aa1940dd8fe32be3d75008f457879b8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>