CMake: Fix SBOM FilesAnalyzed and PackageVerificationCode handling
Packages that don't contain files should have the FilesAnalyzed field
set to false, and should not include a PackageVerificationCode.

Previously we always set FilesAnalyzed to true for any package, and
we only provided a PackageVerificationCode for the top-level repo
package which doesn't actually have files.

The decision whether a package has a file is quite involved, because
it is influenced by the presence of either binary files or custom
files.

To reliably detect this without rewriting a lot of the code, we
postpone the value of FilesAnalyzed to the end of the sbom generation
process, where we can check if any files were added to the package.
This is achieved by setting a package-specific variable to either true
or false, with the final value being evaluated by a configure_file.

To correctly handle PackageVerificationCode, we now collect all file
SHA1s per package, rather than per repo, and compute the verification
code for each package. This is done in a similar fashion to the
FilesAnalyzed field, in a multi-step process, with a final replacement
in configure_file.

The root package doesn't get a PackageVerificationCode anymore because
it doesn't have files.

This fixes the warnings reported by the spdx online validator at
https://tools.spdx.org/app/validate/ which uses the java-based
validator from https://github.com/spdx/tools-java/

Pick-to: 6.9 6.8
Change-Id: I59c4279de6b159a22d6cb36af4b53bc4dea9901b
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 5180b172d9e2256925b2723d30f5e18c8726399d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>