QAbstractItemView: Fix isPersistentEditorOpen for non-persistent editors
isPersistentEditorOpen() returned true if an editor widget existed for
the given index, even when it was not opened as a persistent editor.

Update the implementation to check both that the editor widget exists
and that it is listed in the persistent editor set.

Pick-to: 6.9
Fixes: QTBUG-72333
Change-Id: I380fb51a4e357b5497fb1462e4dcd40c7ead4e13
Reviewed-by: David Faure <david.faure@kdab.com>
(cherry picked from commit c8b7db36a021d53f6a0b76fc31e45421f66c8439)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>