Consider cross compilation when detecting host tools requirements
If Qt is crosscompiled it's not necessary that it always require host
tools to compile the application. That's why the pre-stored assumpion
in Qt6Dependencies.cmake is incorrect. When moving the Qt toolchain
to the target platform, this Qt becomes the native toolchain for this
platform, so the requirement is incorrect since this moment.

The detection should consider the CMAKE_CROSSCOMPILING flag and put
the host tools requirement only in this case. Also if user predefined
QT_HOST_TOOLS it also makes sense to require them and look for the
Qt6HostInfo package.

Change-Id: I4c715eb2592ce79565787645aac172ff60b91493
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 4f44f433329fcc3c81a9de71f9460dee8d962223)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>