tst_QObject: add tests for string-based connect() for invokables
The string-based {dis,}connect() work with methods registered with the
meta-object system, which includes invokables. The SLOT() macro at the
moment works with both slots (Q_SLOTS:) and invokable methods, which may
be considered a bug; only a function marked as a slot should be found by
the SLOT() macro.

Q_INVOKABLE has no effect on a signal, moc will register it as a signal
not just a "method". Extend the tests to show that.

Task-number: QTBUG-139128
Pick-to: 6.9 6.8
Change-Id: Ic62197ab840d5fdeb75ce3da804109d3d969d8a1
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 7f6d53808964140e9606be372883427053e5523b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>