QMacStyle: fall back to draw QProgressBar manually on macOS Tahoe 26
On macOS Tahoe 26, the drawing of a QProgressBar using the mac
style is broken. Nothing is drawn if we continue to use
[NSView drawRect:], [NSView bitmapImageRepForCachingDisplayInRect:]
or [CALayer renderInContext:] (and similar API) for drawing
an NSProgressIndicator.

This patch will therefore implement a fall back approach where
we draw the QProgressBar manually when we detect that the current
version of macOS is running with liquid glass support.

Pick-to: 6.9 6.8 6.5
Fixes: QTBUG-138947
Task-number: QTBUG-138942
Change-Id: Idab450ea1598f7a46fece68c7e3d07f21569cfbb
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit ae6f389942d60959fcae0f5d9f2bb6eadd8efb06)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>