QTest::qWaitFor*: capture the widget/window using QPointer
So we know if they get deleted before the waiting completes. This avoids
a crash if we dereference a dangling pointer or an ABA coincidence in
the case of qWaitForWindowFocused.

There's currently no way to tell QTest::qWaitFor() that the condition
will never become true, so this will result in a timeout.

[ChangeLog][QtTest] Fixed a bug that would cause the qWaitForWindow* and
qWaitForWidget* functions to crash or misbehave if the window or widget
they were told to wait on was destroyed in the process of waiting for it
to be shown/focused/activated/exposed.

Pick-to: 6.8
Task-number: QTBUG-136653
Change-Id: Idaef89fa55e869d6946afffd727653a1fee3770c
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit f0019c51397e75c261777ef2d7c82a96037d9c7f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 62583dac9f3f38769b5740611572822817e6daa9)