tst_QQuaternion: add some more Euler angle tests
In particular, `xonly-opposite` is there to verify that a replacement
of a static const QQuaternion::fromEulerAngles(-90, 0, 0) in QQuick3D
(which is a runtime initialization because fromEulerAngles() isn't
constexpr) with a direct constructor call (which is a compile-time
constant) is equivalent.

The others are for symmetry.

Pick-to: 6.9 6.8 6.5
Change-Id: I20236d7b43680611daa298c5a3266179dd2d2167
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit efc4b92bb00d5eeb338016cece36c4393c54e969)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>