Doc: Add squish testing info to example template
There is a squish-tested-example.qdocinc file which contains squish testing
information for examples.

This patch includes this file into the app-exampls-template file as an optional
section. It should be added if the example was tested with squish.

The README was also updated to include this information.

Fixes: QTBUG-139284
Change-Id: I69fde08cab19fb81897de2066e553a9d4a51d823
Reviewed-by: Jerome Pasion <jerome.pasion@qt.io>
(cherry picked from commit f6435378d43d97e1e26b8ad5e5fc5c560fbc077a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>