Android: Set name-matched class names for Slider and ScrollBar for TalkBack
Seekbar is not the ideal class name for ScrollBar and Slider. It changes
the behavior of these Element, so that the scroll actions are not
properly delivered to them by TalkBack. We use the Android class name
that best match the QAccessible role name. We accept, that TalkBack has
no implementation for them and will use the default behavior of
Android.view.

[ChangeLog][Accessibility][Android] Use different Android class names
for Slider and Scroller.

Pick-to: 6.9 6.8
Change-Id: I03d9ce2bd051ff2c330852636d92e47c6d971884
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit b682d31d7e58cd74bb32df2fc984325f0825020e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>