QUnicodeTools: fix attibute location on properties() functions
The Q_DECL_CONST_FUNCTION needs to be on the declaration to have any
effect on callers, but it was only on the (out-of-line) definition.

Amends 2fe90a61bdf16bb1a08817ba544e2309b524a052.

As a drive-by, also remove the export macros from the definitions;
they, too, are only needed on the declaration.

Pick-to: 6.9 6.8 6.5
Change-Id: Id69b58c50440b8b835f7be7ba873927d07b11219
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit 1b7418e49448413fb78de97743ac5b230e4f3194)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>